changed unsigned int to size_t

+++ philo_one/execution/fork_actions.c
@@ -6,13 +6,13 @@
 /*   By: gbouwen <gbouwen@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/03/03 15:11:18 by gbouwen       #+#    #+#                 */
-/*   Updated: 2021/03/10 11:29:10 by gbouwen       ########   odam.nl         */
+/*   Updated: 2021/03/11 13:03:55 by gbouwen       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../philo_one.h"
 
-void	take_forks(t_philo *philo, unsigned int left, unsigned int right)
+void	take_forks(t_philo *philo, size_t left, size_t right)
 {
 	if (philo->id % 2 == 1)
 	{
@@ -37,7 +37,7 @@ void	take_forks(t_philo *philo, unsigned int left, unsigned int right)
 	philo->status = EATING;
 }
 
-void	drop_forks(t_philo *philo, unsigned int left, unsigned int right)
+void	drop_forks(t_philo *philo, size_t left, size_t right)
 {
 	pthread_mutex_unlock(&philo->data->forks[left]);
 	pthread_mutex_unlock(&philo->data->forks[right]);
